Corrigindo o método de divisão

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
 		
 		int soma = soma(a, b);
 		int subtracao = subtracao(a, b);
-		int divisao = divisao(a, b);
+		double divisao = divisao(a, b);
 		int multiplicacao = multiplicacao(a, b);
 		
 		System.out.println("Soma: " + soma);
@@ -35,7 +35,7 @@ public class Calculadora {
 	public static int multiplicacao(int a, int b) {
 		return a * b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(int a, int b) {
 		return a / b;
 	}
 }
